convert CurrentTick to int before comparison in playtimer_Tick

+++ 8beatMap/Form1.cs
@@ -514,7 +514,7 @@ namespace _8beatMap
         {
             PositionPanel(ConvertTimeToTicks(MusicFileReader.CurrentTime));
 
-            if (CurrentTick != LastTick)
+            if ((int)CurrentTick != LastTick)
             {
                 int ltick = LastTick;
                 LastTick = (int)CurrentTick;
